Allow Top Instructions scrolling with Blockly Blocks
Previously, the read-only BlockSpace we create to hold block hints would
capture the scroll event when the mouse was in its area, preventing the
user from being able to scroll the Top Instructions.

This adds an optional "noScrolling" flag to the creation of a new
read-only blockspace, that prevents it from capturing scrolls.